requirements.txt
requests
google-generativeai
urllib3